js: converted teams files to vue format
