--- app/views/teams/components/_form.html.slim
@@ -1,22 +1,0 @@
-= form_for :team, url: teams_path, html: {id: 'new-team-form', class: 'form-horizontal collapse', role: 'form', novalidate: true, ref: "form", "@submit.prevent" => "onSubmit"} do |f|
-  .form-group :class=="{ 'has-error': $v.team.name.$error }"
-    = f.label :name, {class: 'control-label col-md-2'}
-    .col-md-7
-      = f.text_field(:name, class: 'form-control', required: true, placeholder: "New team's name".html_safe, "@input" => "$v.team.name.$touch()", "v-model.trim" => "team.name")
-      span.help-block
-        span v-if="!$v.team.name.required"
-          | Name can't be blank
-        span v-if="!$v.team.name.available"
-          | Name is reserved or has already been taken
-  - if current_user.admin?
-    .form-group
-      = f.label :owner, {class: 'control-label col-md-2'}
-      .col-md-7
-        = f.select(:owner_id, options_for_select(User.enabled.map { |u| [u.username, u.id] }, current_user.id), {}, { class: 'form-control', "v-model" => "team.owner_id"})
-  .form-group
-    = f.label :description, {class: 'control-label col-md-2'}
-    .col-md-7
-      = f.text_area(:description, class: 'form-control fixed-size', required: false, placeholder: "A short description of your team", "v-model" => "team.description")
-  .form-group
-    .col-md-offset-2.col-md-7
-      = f.submit('Add', class: 'btn btn-primary', ":disabled" => "$v.$invalid")
