--- app/views/teams/components/_table.html.slim
@@ -1,35 +1,0 @@
-div
-  table.table.table-striped.table-hover :class=="{'table-sortable': sortable}"
-    colgroup
-      col.col-5
-      col.col-30
-      col.col-25
-      col.col-20
-      col.col-20
-    thead
-      tr
-        th
-        th @click=="sort('name')"
-          i.fa.fa-fw.fa-sort{
-            :class=="{
-              'fa-sort-amount-asc': sorting.by === 'name' && sorting.asc,
-              'fa-sort-amount-desc': sorting.by === 'name' && !sorting.asc,
-            }"
-          }
-          | Team
-
-        th @click=="sort('role')"
-          i.fa.fa-fw.fa-sort{
-            :class=="{
-              'fa-sort-amount-asc': sorting.by === 'role' && sorting.asc,
-              'fa-sort-amount-desc': sorting.by === 'role' && !sorting.asc,
-            }"
-          }
-          | Role
-        th Number of members
-        th Number of namespaces
-    tbody
-      <team-table-row v-for="team in filteredTeams" :key="team.id" :team="team" :teams-path="teamsPath"></team-table-row>
-
-  <table-pagination :total.sync="teams.length" :current-page="currentPage" :itens-per-page.sync="limit" @update="updateCurrentPage"></table-pagination>
-
