--- app/views/teams/components/_table_row.html.slim
@@ -1,12 +1,0 @@
-tr :class="scopeClass"
-  td.table-icon
-    i class="fa fa-lg" :class="teamIcon"
-  td
-    a :href="teamPath"
-      | {{ team.name }}
-  td
-    | {{ team.role }}
-  td
-    | {{ team.users_count }}
-  td
-    | {{ team.namespaces_count }}
